chore: move util to plotting utils
